Update Navbar.jsx
Update "Link" path to "Bookmarks" page.

A class is added to the <svg> element when the corresponding route matches the current location. This is accomplished using the "useLocation" hook provided by "react-router-dom".

A class conditionally based on whether the location.pathname matches the path of each link has been added; If the condition is true, it adds the class "active" to the <svg> element.